More minor changeS

+++ Project_Files/WindowsApp/Operator911_App/src/operator/Models/PersonModel.java
@@ -59,7 +59,7 @@ public class PersonModel {
      * @return Returns the latitude of the person
      */
     public String getLatitude() {
-        if(Integer.valueOf(latitude)!=0)
+        if(Double.valueOf(latitude)!=0)
             return latitude;
         else
             return "42.026868";
@@ -70,7 +70,7 @@ public class PersonModel {
      * @return Returns the longitude of the person
      */
     public String getLongitude() {
-        if (Integer.valueOf(longitude) != 0)
+        if (Double.valueOf(longitude) != 0)
             return longitude;
         else
             return "-93.644317";
